requirements.txt
Flask==3.0.0
Jinja2==3.1.2
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==3.0.1
